LayoutTests/http/wpt/credential-management/credentialscontainer-store-basics.https.html

<!DOCTYPE html>
<title>Credential Management API: store() basics.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    function asciiToUint8Array(str)
    {
        var chars = [];
        for (var i = 0; i < str.length; ++i)
            chars.push(str.charCodeAt(i));
        return new Uint8Array(chars);
    }

    promise_test(async function(t) {
        const options = {
            publicKey: {
                rp: {
                    name: "localhost",
                },
                user: {
                    name: "John Appleseed",
                    id: asciiToUint8Array("123456"),
                    displayName: "Appleseed",
                },
                challenge: asciiToUint8Array("123456"),
                pubKeyCredParams: [{ type: "public-key", alg: -7 }],
            }
        };

        const credential = await navigator.credentials.create(options);
        return promise_rejects(t, "NotSupportedError",
            navigator.credentials.store(credential));
    }, "navigator.credentials.store().");
</script>
