LayoutTests/http/wpt/credential-management/credentialscontainer-store-basics.https-expected.txt

CONSOLE MESSAGE: User gesture is not detected. To use the WebAuthn API, call 'navigator.credentials.create' within user activated events.

PASS navigator.credentials.store().
